% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare.R
\name{compareplot}
\alias{compareplot}
\title{Create Plot to Compare different Time Series}
\usage{
compareplot(df, var = "year", type = "continue")
}
\arguments{
\item{df}{a data.frame with first col times, second col values.}
}
\value{
Invisible NULL
}
\description{
Create multiple time series on the same scale on the same plot, even the
 components of the times (year or month or date) are different.
}

