% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util-dates.R
\name{create_times}
\alias{create_times}
\title{Create times}
\usage{
create_times(start, end, by, n, tz = "UTC", method = c("sequence",
  "random")[1], ...)
}
\arguments{
\item{start}{start time index}

\item{end}{end time index}

\item{by}{if method is sequence, seq() argument by}

\item{n}{if method is random, vector of length n dates to create}

\item{method}{partial matching methods to be used}
}
\value{
a vector of dates
}
\description{
Create times with different methods
}
\examples{
t1 <- create_times("2015-01-01 00:00:00", "2015-01-01 23:59:59", by=1, n=10)
}

