% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timechunk.R
\name{getborders}
\alias{getborders}
\title{Return borders of time chunks}
\usage{
getborders(df, max_gap = 600, min_len = 10, ind_time = F)
}
\arguments{
\item{df}{a data frame with one column of "time" and of class POSIXct}

\item{max_gap}{maximum time gap allowed to be counted as one chunk [s]}

\item{min_len}{minimum length of a chunk}

\item{ind_time}{a logical value, if TRUE, time index will be returned}
}
\value{
a data frame with two columns:
 - start: start time index of each chunk
 - end: end time index of each chunk
}
\description{
Return the start and end time index of each chunk
}
\examples{
\dontrun{
 br1 <- getborders(df)
 br2 <- getborders(df, max_gap = 100, min_len = 0, ind_time = T)
}
}

