% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timechunk.R
\name{getchunk}
\alias{getchunk}
\title{Return consecutive time chunks}
\usage{
getchunk(df, max_gap = 600, min_len = 10)
}
\arguments{
\item{df}{a data frame with a column named 'time' and of class 'POSIXct'}

\item{max_gap}{maximum time gap allowed to be counted as one chunk [s]}

\item{min_len}{minimum length of a chunk}
}
\value{
a list of splitted chunks
}
\description{
Split data frame into consecutive chunks based on the column time and two
 pre-defined chunk-related thresholds.
}
\examples{
\dontrun{
 chn1 <- getchunk(df)
 chn2 <- getchunk(df, max_gap = 100, min_len = 0)
}
}

