% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getfileext}
\alias{getfileext}
\title{Get File Extention}
\usage{
getfileext(path)
}
\arguments{
\item{path}{character of length 1}
}
\value{
character of file extension
}
\description{
Get File Extention
}
\examples{
fileext <- getfileext("example.R")
}

