% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{statusplot}
\alias{statusplot}
\title{Create Process Status Plot}
\usage{
statusplot(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
Invisible NULL
}
\description{
Create process status plot, with each bar indicating each process, and each
color indicating different categories of status.
}
\examples{
statusplot(data.frame(name = c("name1", "name2", "name3"),
 start = c("20150101", "20150201", "20150301"),
 end = c("20150110", "20150220", "20150330"),
 flag = c("good", "med", "bad")))
}

